+++ app/src/utils/gol-pattern-store.ts
